+++ frontend/openchat-agent/src/services/localUserIndex/candid/idl.js
@@ -420,6 +420,7 @@ export const idlFactory = ({ IDL }) => {
     'memo' : IDL.Opt(IDL.Nat64),
     'ledger' : CanisterId,
     'amount' : Tokens,
+    'symbol' : IDL.Text,
   });
   const SnsPendingCryptoTransaction = IDL.Record({
     'to' : Icrc1Account,
@@ -429,6 +430,7 @@ export const idlFactory = ({ IDL }) => {
     'memo' : IDL.Opt(IDL.Nat64),
     'ledger' : CanisterId,
     'amount' : Tokens,
+    'symbol' : IDL.Text,
   });
   const Icrc1PendingCryptoTransaction = IDL.Record({
     'to' : Icrc1Account,
@@ -438,6 +440,7 @@ export const idlFactory = ({ IDL }) => {
     'memo' : IDL.Opt(Memo),
     'ledger' : CanisterId,
     'amount' : IDL.Nat,
+    'symbol' : IDL.Text,
   });
   const PendingCryptoTransaction = IDL.Variant({
     'NNS' : NnsPendingCryptoTransaction,