+++ frontend/openchat-agent/src/services/localUserIndex/candid/types.d.ts
@@ -635,6 +635,7 @@ export interface Icrc1PendingCryptoTransaction {
   'memo' : [] | [Memo],
   'ledger' : CanisterId,
   'amount' : bigint,
+  'symbol' : string,
 }
 export interface ImageContent {
   'height' : number,
@@ -886,6 +887,7 @@ export interface NnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface NnsProposal {
   'id' : ProposalId,
@@ -1154,6 +1156,7 @@ export interface SnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface SnsProposal {
   'id' : ProposalId,