+++ frontend/openchat-agent/src/services/user/candid/types.d.ts
@@ -862,6 +862,7 @@ export interface Icrc1PendingCryptoTransaction {
   'memo' : [] | [Memo],
   'ledger' : CanisterId,
   'amount' : bigint,
+  'symbol' : string,
 }
 export interface ImageContent {
   'height' : number,
@@ -1090,6 +1091,7 @@ export interface NnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface NnsProposal {
   'id' : ProposalId,
@@ -1496,6 +1498,7 @@ export interface SnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface SnsProposal {
   'id' : ProposalId,