+++ frontend/openchat-agent/src/services/userIndex/candid/types.d.ts
@@ -673,6 +673,7 @@ export interface Icrc1PendingCryptoTransaction {
   'memo' : [] | [Memo],
   'ledger' : CanisterId,
   'amount' : bigint,
+  'symbol' : string,
 }
 export interface ImageContent {
   'height' : number,
@@ -831,6 +832,7 @@ export interface NnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface NnsProposal {
   'id' : ProposalId,
@@ -1151,6 +1153,7 @@ export interface SnsPendingCryptoTransaction {
   'memo' : [] | [bigint],
   'ledger' : CanisterId,
   'amount' : Tokens,
+  'symbol' : string,
 }
 export interface SnsProposal {
   'id' : ProposalId,