+++ frontend/openchat-shared/src/domain/chat/chat.ts
@@ -70,6 +70,7 @@ export type CryptocurrencyDeposit = {
 export type PendingCryptocurrencyWithdrawal = {
     kind: "pending";
     ledger: string;
+    symbol: string;
     token: Cryptocurrency;
     to: string;
     amountE8s: bigint;
@@ -124,6 +125,7 @@ export type CompletedCryptocurrencyTransfer = {
 export type PendingCryptocurrencyTransfer = {
     kind: "pending";
     ledger: string;
+    symbol: string;
     token: Cryptocurrency;
     recipient: string;
     amountE8s: bigint;